requirements.txt
aniso8601==1.2.0
blinker==1.4
click==6.7
dominate==2.3.1
Flask==0.12
Flask-Cors==3.0.2
Flask-JWT==0.3.2
Flask-Mail==0.9.1
Flask-PyMongo==0.4.1
Flask-Redis==0.3.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-Uploads==0.2.1
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
PyJWT==1.4.2
pymongo==3.4.0
python-dateutil==2.6.0
pytz==2016.10
redis==2.10.5
six==1.10.0
visitor==0.1.3
Werkzeug==0.11.15
WTForms==2.1
